+++ VadereState/src/org/vadere/state/scenario/Pedestrian.java
@@ -4,9 +4,10 @@ import org.vadere.state.attributes.scenario.AttributesAgent;
 import org.vadere.state.behavior.SalientBehavior;
 import org.vadere.state.events.types.Event;
 import org.vadere.state.simulation.FootStep;
-import org.vadere.state.simulation.LastFootSteps;
+import org.vadere.state.simulation.FootstepHistory;
 import org.vadere.state.simulation.VTrajectory;
 import org.vadere.state.types.ScenarioElementType;
+import org.vadere.util.geometry.shapes.VPoint;
 
 import java.util.HashMap;
 import java.util.LinkedList;
@@ -17,6 +18,7 @@ public class Pedestrian extends Agent {
 
 	// Constants
 	public static double PEDESTRIAN_MAX_SPEED_METER_PER_SECOND = 12.0;
+	public static final double INVALID_NEXT_EVENT_TIME = -1.0;
 
 	// Variables
 	private int idAsTarget; // TODO should actually be an attribute or a member of a subclass
@@ -26,13 +28,15 @@ public class Pedestrian extends Agent {
 	private SalientBehavior salientBehavior;
 	private LinkedList<Integer> groupIds; // TODO should actually be an attribute or a member of a subclass
 	/**
-	 * Footsteps is a list of foot steps a pedestrian made during the duration of one time step.
+	 * trajectory is a list of foot steps a pedestrian made during the duration of one time step.
 	 * For all non event driven models this is exactly one foot step. For the event driven update
 	 * one pedestrian can move multiple times during one time step. To save memory the list of foot steps
 	 * will be cleared after each completion of a time step. The output processor <tt>PedestrianStrideProcessor</tt>
 	 * can write out those foot steps.
 	 */
 	private VTrajectory trajectory;
+	/** This list stores the last n footsteps. I.e., this list is NOT cleared after each simulation loop like "trajectory" variable. */
+	private transient FootstepHistory footstepHistory;
 
 	private LinkedList<Integer> groupSizes;
 	private Map<Class<? extends ModelPedestrian>, ModelPedestrian> modelPedestrianMap;
@@ -59,7 +63,8 @@ public class Pedestrian extends Agent {
 		groupIds = new LinkedList<>();
 		groupSizes = new LinkedList<>();
 		modelPedestrianMap = new HashMap<>();
-		trajectory = new VTrajectory(attributesAgent.getFootStepsToStore());
+		trajectory = new VTrajectory();
+		footstepHistory = new FootstepHistory(attributesAgent.getFootstepHistorySize());
 	}
 
 	private Pedestrian(Pedestrian other) {
@@ -81,6 +86,7 @@ public class Pedestrian extends Agent {
 
 		trajectory = new VTrajectory();
 		trajectory = other.trajectory;
+		footstepHistory = other.footstepHistory;
 	}
 
 	// Getter
@@ -105,10 +111,33 @@ public class Pedestrian extends Agent {
 		return ScenarioElementType.PEDESTRIAN;
 	}
 
-	// TODO Rename "getFootSteps()" to "getTrajectory()".
-	public VTrajectory getFootSteps() {
+	public VTrajectory getTrajectory() {
 		return trajectory;
 	}
+
+	public FootstepHistory getFootstepHistory() {
+		return footstepHistory;
+	}
+
+    public VPoint getInterpolatedFootStepPosition(double time){
+        if(this.footstepHistory.getCapacity() <= 0){
+            throw new IllegalArgumentException("Cannot interpolate foot steps if there is no capacity (larger than zero) " +
+                    "for storing foot steps (see 'scenario.attributesPedestrian.footStepsToStore' field)");
+        }
+
+        FootStep currentFootStep = this.footstepHistory.getYoungestFootStep();
+
+        if(currentFootStep == null){
+            return getPosition();
+        }else{
+            if(time > currentFootStep.getEndTime()){
+                // This happens for example if a pedestrian is waiting (see Events)
+                return currentFootStep.getEnd();
+            }else{
+                return FootStep.interpolateFootStep(currentFootStep, time);
+            }
+        }
+    }
 
 	// Setter
 	public void setIdAsTarget(int id) { this.idAsTarget = id; }
@@ -140,8 +169,17 @@ public class Pedestrian extends Agent {
 		groupSizes.add(size);
 	}
 
+	public void addFootStepToTrajectory(FootStep footStep){
+		this.trajectory = this.trajectory.add(footStep);
+	}
+
 	public void clearFootSteps() {
-		trajectory.clear();
+		// getLast() is always the most recent (made sure in VTrajectory.add)
+		// This statement is for security and should mostly have no effect (only if someone did not use method
+		// "addFootStepToTrajectory" to add another foot step to the trajectory)
+		if(!trajectory.isEmpty()){
+			trajectory.clear();
+		}
 	}
 
 	// Overridden Methods