+++ VadereState/src/org/vadere/state/simulation/FootStep.java
@@ -42,7 +42,9 @@ public final class FootStep {
 	 * @param endTime   end time of the foot step
 	 */
 	public FootStep(@NotNull final VPoint start, @NotNull final VPoint end, final double startTime, final double endTime) {
-		this.start = start;
+        assert startTime <= endTime;
+
+	    this.start = start;
 		this.end = end;
 		this.startTime = startTime;
 		this.endTime = endTime;
@@ -122,9 +124,46 @@ public final class FootStep {
 		return intersectionTime;
 	}
 
+	static public VPoint interpolateFootStep(final FootStep footStep, final double time){
+		final double startTime = footStep.getStartTime();
+		final double endTime = footStep.getEndTime();
+		final double duration = footStep.duration();
+
+		if(startTime > time || endTime < time || startTime < 0 ){
+			throw new IllegalArgumentException("Requested time " + time + " outside of valid time " +
+					"region (no extrapolation!). Value outside of FootStep " +
+					"[start=" + startTime + ", end=" + endTime + "] time or smaller than zero. " +
+					"\n FootStep=" + footStep.toString());
+		}
+
+		VPoint interpolationResult;
+
+		if(duration < 1E-14){
+			// to avoid problems with division by "very small number", simply return the start point
+			interpolationResult = footStep.getStart();
+		}else{
+			double linearTime = (time - startTime) / duration;
+
+			VPoint diffPoint = footStep.getEnd().subtract(footStep.getStart());
+
+			diffPoint.x = diffPoint.x * linearTime;
+			diffPoint.y = diffPoint.y * linearTime;
+
+			interpolationResult = footStep.getStart().add(diffPoint);
+		}
+
+		return interpolationResult;
+	}
+
+
 	@Override
 	public String toString() {
-		return start + "->" + end;
+		return "start=" + start + "(t=" + startTime + ")" + "->" + "end=" + end + "(t=" + endTime + ")";
+	}
+
+	@Override
+	public FootStep clone(){
+		return new FootStep(this.start.clone(), this.end.clone(), this.startTime, this.endTime);
 	}
 
 	public String[] getValueString(){