+++ VadereState/src/org/vadere/state/simulation/FootstepHistory.java
@@ -1,21 +1,24 @@
 package org.vadere.state.simulation;
 
+import org.jetbrains.annotations.Nullable;
+
 import java.util.ArrayList;
 import java.util.stream.Collectors;
 
 /**
  * Store the last foot steps of an agent to calculate the agent's average speed during simulation.
- *
- * TODO: Maybe, rename to "FootStepHistory".
  */
-public class LastFootSteps {
+public class FootstepHistory {
 
     // Variables
     private int capacity;
     private ArrayList<FootStep> footSteps;
 
     // Constructors
-    public LastFootSteps(int capacity) {
+    public FootstepHistory() {
+        this(10);
+    }
+    public FootstepHistory(int capacity) {
         this.capacity = capacity;
         this.footSteps = new ArrayList<>(capacity);
     }
@@ -61,6 +64,7 @@ public class LastFootSteps {
         return oldestFootStep;
     }
 
+    @Nullable
     public FootStep getYoungestFootStep() {
         FootStep youngestFootStep = null;