+++ VadereState/tests/org/vadere/state/attributes/TestAttributesODEModel.java
@@ -9,6 +9,8 @@ import org.vadere.state.attributes.models.AttributesODEIntegrator;
 import org.vadere.state.types.IntegratorType;
 import org.vadere.state.util.StateJsonConverter;
 
+import java.io.IOException;
+
 public class TestAttributesODEModel {
 
 	private static final double delta = 1e-8;
@@ -21,22 +23,20 @@ public class TestAttributesODEModel {
 	@Before
 	public void setUp() {
 		store = "{" + "\"solverType\" : \"CLASSICAL_RK4\","
-				+ "\"toleranceAbsolute\" : " + new Double(1e-5).toString() + "," + "\"toleranceRelative\" : "
-				+ new Double(1e-5).toString() + "," + "\"stepSizeMin\" : " + new Double(1e-5).toString() + ","
-				+ "\"stepSizeMax\" : " + new Double(1e-5).toString() + "}";
+				+ "\"toleranceAbsolute\" : " + Double.toString(1e-5) + "," + "\"toleranceRelative\" : "
+				+ Double.toString(1e-5) + "," + "\"stepSizeMin\" : " + Double.toString(1e-5) + ","
+				+ "\"stepSizeMax\" : " + Double.toString(1e-5) + "}";
 	}
 
 	/**
 	 * Test method for
-	 * {@link org.vadere.state.attributes.models.AttributesODEIntegrator#AttributesODEModel(java.util.Map)}
+	 * {@link org.vadere.state.attributes.models.AttributesODEIntegrator(java.util.Map)}
 	 * . Asserts that creating an {@link AttributesODEIntegrator} with the given
 	 * store sets the correct instance variables.
-	 * 
-	 * @throws IllegalAccessException
 	 * @throws IllegalArgumentException
 	 */
 	@Test
-	public void testAttributesODEModel() throws IllegalArgumentException, IllegalAccessException {
+	public void testAttributesODEModel() throws IllegalArgumentException, IOException {
 		// correct case
 		attributesODEModel = StateJsonConverter.deserializeObjectFromJson(store, AttributesODEIntegrator.class);
 		assertArrayEquals(new double[] {1e-5}, new double[] {attributesODEModel.getToleranceAbsolute()}, delta);