+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 ### Added
 
 ### Changed
+- `PedestrianFootStepProcessor` interpolates the pedestrian's foot step to obtain a more precise position.
+
+# v1.4 (2019-09-05)
+
+### Changed 
+
+- Renamed "footStepsToStore" to "footstepHistorySize" in all scenario files.
 
 ## v1.3 (2019-07-31)